Moved `BehaviourSummary` classes in `dto` packages
These are `data` classes but they are specifically dto (data transfer objects)
classes because they're used to serialise data to the API client.

Moved to `dto.*` package for consistency with other similar classes.
